REFACTORING. Created 'PageNav' component and use it on the 'Header' component
